table tooltips for clipped cells
