+++ src/framework/styles.css
@@ -23,3 +23,11 @@
 ::-webkit-scrollbar-thumb:hover {
   background: theme('colors.grey2');
 }
+
+.tippy-box[data-theme~='wordcloud'] {
+  background-color: #333b;
+  backdrop-filter: blur(5px);
+  color: white;
+  padding: 0.5rem;
+  border-radius: 0.5rem;
+}
