+++ src/framework/visualisation/react/ui/elements/figure.tsx
@@ -2,36 +2,19 @@ import { TableWithContext } from '../../../../types/elements'
 
 import TextBundle from '../../../../text_bundle'
 import { Translator } from '../../../../translator'
-import {
-  VisualizationType,
-  VisualizationData,
-  AxisSettings,
-  TickerFormat
-} from '../../../../types/visualizations'
+import { VisualizationType, VisualizationData } from '../../../../types/visualizations'
 
-import React, { useMemo, useState } from 'react'
+import { useMemo } from 'react'
 
 import { ReactFactoryContext } from '../../factory'
-import {
-  ResponsiveContainer,
-  LineChart,
-  Line,
-  XAxis,
-  YAxis,
-  Tooltip,
-  Legend,
-  BarChart,
-  Bar,
-  AreaChart,
-  Area
-} from 'recharts'
 
 import useVisualizationData from '../hooks/useVisualizationData'
 import { Title6 } from './text'
 
 import Lottie from 'lottie-react'
 import spinnerDark from '../../../../../assets/lottie/spinner-dark.json'
-import ReactWordcloud, { Options } from 'react-wordcloud'
+import RechartsGraph from './figures/recharts_graph'
+import Wordcloud from './figures/wordcloud'
 
 type Props = VisualizationProps & ReactFactoryContext
 
@@ -71,13 +54,15 @@ export const Figure = ({
 
   const minHeight = visualization.height ? visualization.height + 'px' : `20rem`
 
-  console.log(visualizationData)
   return (
-    <div className="flex flex-col overflow-hidden ">
+    <div className="flex flex-col overflow-hidden">
       <Title6 text={title} margin="mt-2 mb-4" />
-      <div className={`relative z-50 flex `} style={{ flex: `1 1 ${minHeight}`, minHeight }}>
+      <div
+        className={`relative z-50 flex max-w-full`}
+        style={{ flex: `1 1 ${minHeight}`, minHeight }}
+      >
         {visualizationData?.data.length === 0 ? (
-          <div className="m-auto  font-bodybold text-4xl text-grey2">{noDataMsg}</div>
+          <div className="m-auto font-bodybold text-4xl text-grey2 ">{noDataMsg}</div>
         ) : (
           <RenderVisualization visualizationData={visualizationData} />
         )}
@@ -94,166 +79,11 @@ const RenderVisualization = ({
   if (!visualizationData) return null
 
   if (['line', 'bar', 'area'].includes(visualizationData.type))
-    return <RenderRechartsGraph visualizationData={visualizationData} />
+    return <RechartsGraph visualizationData={visualizationData} />
 
   if (visualizationData.type === 'wordcloud')
-    return <RenderWordcloud visualizationData={visualizationData} />
+    return <Wordcloud visualizationData={visualizationData} />
   return null
-}
-
-interface RenderVisualizationProps {
-  visualizationData: VisualizationData
-}
-
-const RenderRechartsGraph = ({
-  visualizationData
-}: RenderVisualizationProps): JSX.Element | null => {
-  function tooltip() {
-    return (
-      <Tooltip
-        allowEscapeViewBox={{ x: false, y: false }}
-        labelStyle={{ marginBottom: '0.5rem' }}
-        contentStyle={{
-          fontSize: '0.8rem',
-          lineHeight: '0.8rem',
-          background: '#fff8',
-          backdropFilter: 'blur(3px)'
-        }}
-      />
-    )
-  }
-
-  function axes(minTickGap: number) {
-    if (!visualizationData) return null
-    const secondary =
-      Object.values(visualizationData.yKeys).findIndex((yKey: AxisSettings) => yKey.secondAxis) !==
-      -1
-    const { tickFormatter, tickFormatter2 } = getTickFormatters(
-      Object.values(visualizationData.yKeys)
-    )
-
-    return (
-      <>
-        <XAxis dataKey={visualizationData.xKey.label} minTickGap={minTickGap} />
-        <YAxis yAxisId="left" tickFormatter={tickFormatter} />
-        {secondary && <YAxis yAxisId="right" orientation="right" tickFormatter={tickFormatter2} />}
-      </>
-    )
-  }
-
-  function legend() {
-    return (
-      <Legend
-        margin={{ left: 10 }}
-        align="right"
-        verticalAlign="top"
-        iconType="plainline"
-        wrapperStyle={{ fontSize: '0.8rem' }}
-      />
-    )
-  }
-
-  let chart: JSX.Element | null = null
-
-  if (visualizationData.type === 'line') {
-    chart = (
-      <LineChart data={visualizationData.data}>
-        {axes(20)}
-        {tooltip()}
-        {legend()}
-        {Object.values(visualizationData.yKeys).map((yKey: AxisSettings, i: number) => {
-          const { color, dash } = getLineStyle(i)
-          return (
-            <Line
-              key={yKey.label}
-              yAxisId={yKey.secondAxis ? 'right' : 'left'}
-              type="monotone"
-              dataKey={yKey.label}
-              dot={false}
-              strokeWidth={2}
-              stroke={color}
-              strokeDasharray={dash}
-            />
-          )
-        })}
-      </LineChart>
-    )
-  }
-
-  if (visualizationData.type === 'bar') {
-    chart = (
-      <BarChart data={visualizationData.data}>
-        {axes(0)}
-        {tooltip()}
-        {legend()}
-        {Object.values(visualizationData.yKeys).map((yKey: AxisSettings, i: number) => {
-          const { color, dash } = getLineStyle(i)
-          return (
-            <Bar
-              key={yKey.label}
-              yAxisId={yKey.secondAxis ? 'right' : 'left'}
-              dataKey={yKey.label}
-              fill={color}
-            />
-          )
-        })}
-      </BarChart>
-    )
-  }
-
-  if (visualizationData.type === 'area') {
-    chart = (
-      <AreaChart data={visualizationData.data}>
-        {axes(20)}
-        {tooltip()}
-        {legend()}
-        {Object.values(visualizationData.yKeys).map((yKey: AxisSettings, i: number) => {
-          const { color, dash } = getLineStyle(i)
-          return (
-            <Area
-              key={yKey.label}
-              yAxisId={yKey.secondAxis ? 'right' : 'left'}
-              dataKey={yKey.label}
-              fill={color}
-            />
-          )
-        })}
-      </AreaChart>
-    )
-  }
-
-  if (!chart) return null
-  return (
-    <ResponsiveContainer width="100%" height="100%">
-      {chart}
-    </ResponsiveContainer>
-  )
-}
-
-function RenderWordcloud({ visualizationData }: RenderVisualizationProps): JSX.Element | null {
-  const [options] = useState<any>({
-    rotations: 2,
-    rotationAngles: [0],
-    scale: 'sqrt',
-    fontSizes: [10, 50],
-    enableTooltip: true
-  })
-
-  const words = useMemo(() => {
-    const words = []
-    for (let row of visualizationData.data) {
-      const text = row[visualizationData?.xKey.label] as string
-      let value = 0
-      for (let yKey of Object.values(visualizationData.yKeys)) {
-        value += Number(row[yKey.label]) || 0
-      }
-      words.push({ text, value })
-    }
-
-    return words.sort((a, b) => b.value - a.value).slice(0, 50)
-  }, [visualizationData])
-
-  return <ReactWordcloud words={words} options={options} />
 }
 
 function prepareCopy(locale: string): Record<string, string> {
@@ -268,37 +98,3 @@ const noDataMsg = new TextBundle().add('en', 'No data').add('nl', 'Geen data')
 const errorMsg = new TextBundle()
   .add('en', 'Could not create visualization')
   .add('nl', 'Kon visualisatie niet maken')
-
-function getLineStyle(index: number): { color: string; dash: string } {
-  const COLORS = ['#4272EF', '#FF5E5E', '#FFCF60', '#1E3FCC', '#CC3F3F', '#CC9F3F']
-  const DASHES = ['1', '5 5', '10 10', '5 5 10 10']
-
-  const cell = index % (COLORS.length * DASHES.length)
-  const row = index % COLORS.length
-  const column = Math.floor(cell / COLORS.length)
-
-  return { color: COLORS[row], dash: DASHES[column] }
-}
-
-function getTickFormatters(yKeys: AxisSettings[]) {
-  let tickerFormat: TickerFormat | undefined = undefined
-  let tickerFormat2: TickerFormat | undefined = undefined
-  for (let yKey of yKeys) {
-    if (!yKey.secondAxis) {
-      if (!tickerFormat) tickerFormat = yKey.tickerFormat
-      if (tickerFormat !== yKey.tickerFormat) tickerFormat = 'default'
-    } else {
-      if (!tickerFormat2) tickerFormat2 = yKey.tickerFormat
-      if (tickerFormat2 !== yKey.tickerFormat) tickerFormat2 = 'default'
-    }
-  }
-
-  const tickFormatter = getTickFormatter(tickerFormat || 'default')
-  const tickFormatter2 = getTickFormatter(tickerFormat2 || 'default')
-  return { tickFormatter, tickFormatter2 }
-}
-
-function getTickFormatter(format: TickerFormat) {
-  if (format === 'percent') return (value: number) => `${value}%`
-  return undefined
-}
